requirements.txt
ctranslate2>=4,<5
gradio>=6.10,<7
opendataloader-pdf>=2.1.1,<3
PyMuPDF>=1.27.2,<2
python-dotenv>=1.0.0
sentencepiece>=0.2,<1
transformers>=4.46,<4.57
